fix(CarMoveTest, ValidationUtilsTest, SplitCarTest, WinnerCarTest): 해당 TEST CLASS model 디렉토리 하위 위치 이동
