+++ src/test/java/racingcar/model/CarMoveTest.java
@@ -1,8 +1,7 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.model.CarMove;
 
 import static org.assertj.core.api.Assertions.assertThat;
